Add props
