+++ src/components/AnswerForm/index.jsx
@@ -1,25 +1,6 @@
-import React, { useContext, useState, useEffect } from "react";
-import DataContext from "../../context/dataContext";
+import React from "react";
 
-const AnswerForm = () => {
-  const { questionData } = useContext(DataContext);
-  const [answers, setAnswers] = useState([]);
-  const [score, setScore] = useState(0);
-  const [currentQuestion, setCurrentQuestion] = useState(0);
-
-  const calcScore = () => {
-    if (answers == questionData[currentQuestion]["correct_answer"]) {
-      setScore(score + 1);
-    }
-  };
-
-  const handleAnswer = (e) => {
-    e.preventDefault();
-    setAnswers(e.target.value);
-    setCurrentQuestion(currentQuestion + 1);
-    calcScore();
-  };
-
+const AnswerForm = ({ handleAnswer, currentQuestion, questionData }) => {
   return (
     <>
       {questionData.length
